requirements.txt
Package                 Version
----------------------- ------------
absl-py                 2.3.1
asttokens               3.0.0
comm                    0.2.3
contourpy               1.3.3
cycler                  0.12.1
debugpy                 1.8.17
decorator               5.2.1
executing               2.2.1
filelock                3.19.1
fonttools               4.60.1
fsspec                  2025.9.0
grpcio                  1.76.0
imagecodecs             2026.1.1
imageio                 2.37.0
ipykernel               7.1.0
ipython                 9.6.0
ipython_pygments_lexers 1.1.1
jedi                    0.19.2
Jinja2                  3.1.6
jupyter_client          8.6.3
jupyter_core            5.9.1
kiwisolver              1.4.9
lazy_loader             0.4
lpips                   0.1.4
Markdown                3.10
MarkupSafe              2.1.5
matplotlib              3.10.7
matplotlib-inline       0.2.1
mpmath                  1.3.0
nest-asyncio            1.6.0
networkx                3.5
numpy                   2.3.3
nvidia-cublas           13.0.0.19
nvidia-cuda-cupti       13.0.48
nvidia-cuda-nvrtc       13.0.48
nvidia-cuda-runtime     13.0.48
nvidia-cudnn-cu13       9.13.0.50
nvidia-cufft            12.0.0.15
nvidia-cufile           1.15.0.42
nvidia-curand           10.4.0.35
nvidia-cusolver         12.0.3.29
nvidia-cusparse         12.6.2.49
nvidia-cusparselt-cu13  0.8.0
nvidia-nccl-cu13        2.27.7
nvidia-nvjitlink        13.0.39
nvidia-nvshmem-cu13     3.3.24
nvidia-nvtx             13.0.39
packaging               25.0
pandas                  2.3.3
parso                   0.8.5
pexpect                 4.9.0
pillow                  11.3.0
pip                     24.0
platformdirs            4.5.0
prompt_toolkit          3.0.52
protobuf                6.33.0
psutil                  7.1.2
ptyprocess              0.7.0
pure_eval               0.2.3
Pygments                2.19.2
pyparsing               3.2.5
python-dateutil         2.9.0.post0
pytz                    2025.2
PyYAML                  6.0.3
pyzmq                   27.1.0
scikit-image            0.25.2
scipy                   1.16.3
setuptools              70.2.0
six                     1.17.0
stack-data              0.6.3
sympy                   1.14.0
tensorboard             2.20.0
tensorboard-data-server 0.7.2
tifffile                2025.10.16
torch                   2.9.0+cu130
torchvision             0.24.0+cu130
tornado                 6.5.2
tqdm                    4.67.1
traitlets               5.14.3
triton                  3.5.0
typing_extensions       4.15.0
tzdata                  2025.2
wcwidth                 0.2.14
Werkzeug                3.1.3
